Extra flexible event fix
This change fixes the issue where we were should flexible events for the next week in the current week's view. This is because a UTC to ISO conversion ended up flattening time in a way that caused a fence pole issue with the date value, causing the window to look like 8 days instead of 7. The fix look a bit ugly, but works well,